pyproject.toml
[project]
name = "eye-scroll"
version = "0.1.0"
description = "EyeScroll - 眼球追踪滚动应用 macOS 菜单栏工具"
requires-python = ">=3.11"
dependencies = [
    "mediapipe>=0.10.0",
    "opencv-python>=4.8.0",
    "pyobjc-framework-Quartz>=10.0",
    "pyobjc-framework-Cocoa>=10.0",
    "rumps>=0.4.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0"]

[project.scripts]
eye-scroll = "main:main"
